third_party/blink/web_tests/fast/forms/submit-form-with-dirname-attribute-with-ancestor-dir-attribute-expected.txt

Test that when dirname attribute is specified then it is added in submission body.

Post Comment

Hello
PASS document.location.search.indexOf("rtlAncestor.dir=rtl") != -1 is true
PASS document.location.search.indexOf("autoAncestor.dir=ltr") != -1 is true
PASS successfullyParsed is true

TEST COMPLETE
